Move Probe to root
